fix: programs url
The programs url is constructed from the LMS base url and some path. THe
LMS base url is read from the config, however at the time of the
evaluation, the request to the LMS MFEs API is not done, so the config
value is empty. The constructed url ends up being a relative path (since
the base url is empty), so the full path is pointing to MFE and not LMS.

Fixing this by converting the static value to a function (and all the
places where it's used to a function call), so that the value is
constructed dynamically when necessary, i.e. after the config has been
updated with the values returned by the API call.